+++ cli/src/pcluster3_config_converter/pcluster3_config_converter.py
@@ -296,6 +296,7 @@ class Pcluster3ConfigConverter(object):
                 ("efs_kms_key_id", "KmsKeyId"),
                 ("provisioned_throughput", "ProvisionedThroughput", "getint"),
                 ("throughput_mode", "ThroughputMode"),
+                ("access_point_id", "AccessPointId"),
             ]
             efs_section, efs_dict, _section_label = self.convert_storage_base(
                 "efs", efs_label.strip(), additional_items