+++ cli/tests/pcluster/validators/test_efs_validators.py
@@ -11,7 +11,7 @@
 
 import pytest
 
-from pcluster.validators.efs_validators import EfsMountOptionsValidator
+from pcluster.validators.efs_validators import EfsAccessPointOptionsValidator, EfsMountOptionsValidator
 from tests.pcluster.validators.utils import assert_failure_messages
 
 
@@ -27,8 +27,8 @@ from tests.pcluster.validators.utils import assert_failure_messages
             False,
             True,
             "EFS IAM authorization cannot be enabled when encryption in-transit is disabled. "
-            "Please either disable IAM authorization or enable encryption in-transit for file system "
-            "<name-of-the-file-system>",
+            "Please either disable IAM authorization or enable encryption in-transit "
+            "for file system <name-of-the-file-system>",
         ),
         (
             True,
@@ -45,5 +45,72 @@ from tests.pcluster.validators.utils import assert_failure_messages
 def test_efs_mount_options_validator(encryption_in_transit, iam_authorization, expected_message):
     actual_failures = EfsMountOptionsValidator().execute(
         encryption_in_transit, iam_authorization, "<name-of-the-file-system>"
+    )
+    assert_failure_messages(actual_failures, expected_message)
+
+
+@pytest.mark.parametrize(
+    "access_point_id, file_system_id, expected_message",
+    [
+        (
+            None,
+            None,
+            None,
+        ),
+        (
+            "<access_point_id>",
+            None,
+            "An access point can only be specified when using an existing EFS file system. "
+            "Please either remove the access point id <access_point_id> "
+            "or provide the file system id for the access point",
+        ),
+        (
+            "<access_point_id>",
+            "<file-systemd-id>",
+            None,
+        ),
+        (
+            None,
+            "<file-systemd-id>",
+            None,
+        ),
+    ],
+)
+def test_efs_access_point_with_filesystem_validator(access_point_id, file_system_id, expected_message):
+    actual_failures = EfsAccessPointOptionsValidator().execute(access_point_id, file_system_id, True)
+    assert_failure_messages(actual_failures, expected_message)
+
+
+@pytest.mark.parametrize(
+    "access_point_id, encryption_in_transit, expected_message",
+    [
+        (
+            None,
+            False,
+            None,
+        ),
+        (
+            "<access_point_id>",
+            False,
+            "An access point can only be specified when encryption in transit is enabled. "
+            "Please either remove the access point id <access_point_id> or enable encryption in transit.",
+        ),
+        (
+            "<access_point_id>",
+            True,
+            None,
+        ),
+        (
+            None,
+            True,
+            None,
+        ),
+    ],
+)
+def test_efs_access_point_with_filesystem_encryption_validator(
+    access_point_id, encryption_in_transit, expected_message
+):
+    actual_failures = EfsAccessPointOptionsValidator().execute(
+        access_point_id, "<file-system-id>", encryption_in_transit
     )
     assert_failure_messages(actual_failures, expected_message)